Casting tensor to double before calculating mean/std to avoid precision errors.
For large training sets I experienced wildly incorrect results for the mean without this cast.
(Every channel-mean was the same and far off the correct value).
I suspect that this is caused by precision problems due to the mean being implemented as a running sum
which is then divided by the large set size.